# filmApi/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Film API",
        "description": "API de gestion des films, acteurs, critiques et utilisateurs",
        "version": "1.0.0"
    },
    "paths": {
...
        "/users": {
            "post": {
                "tags": [
                    "Users"
                ],
                "summary": "Créer un utilisateur",
                "operationId": "a1fce870c566eb5fc0cf3178ec5d2fed",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "login": {
                                        "type": "string",
                                        "example": "bdje"
                                    },
                                    "email": {
                                        "type": "string",
                                        "example": "bdje@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "Secret123!"
                                    },
                                    "first_name": {
                                        "type": "string",
                                        "example": "Bernadin"
                                    },
                                    "last_name": {
                                        "type": "string",
                                        "example": "Dje"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Utilisateur créé avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "login": {
                                            "type": "string",
                                            "example": "bdje"
                                        },
                                        "email": {
                                            "type": "string",
                                            "example": "bdje@example.com"
                                        },
                                        "first_name": {
                                            "type": "string",
                                            "example": "Bernadin"
                                        },
                                        "last_name": {
                                            "type": "string",
                                            "example": "Dje"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Erreur de validation"
                    }
                }
            }
        },
        "/users/{id}": {
            "put": {
                "tags": [
                    "Users"
                ],
                "summary": "Mettre à jour un utilisateur",
                "operationId": "ddfe232db1870b82493bf8bd25596d37",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de l'utilisateur",
                        "required": true
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "login": {
                                        "type": "string",
                                        "example": "btra"
                                    },
                                    "email": {
                                        "type": "string",
                                        "example": "btra@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "NewSecret123!"
                                    },
                                    "first_name": {
                                        "type": "string",
                                        "example": "Bi"
                                    },
                                    "last_name": {
                                        "type": "string",
                                        "example": "Tra"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Utilisateur mis à jour",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "login": {
                                            "type": "string",
                                            "example": "btra"
                                        },
                                        "email": {
                                            "type": "string",
                                            "example": "btra@example.com"
                                        },
                                        "first_name": {
                                            "type": "string",
                                            "example": "Bi"
                                        },
                                        "last_name": {
                                            "type": "string",
                                            "example": "Tra"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Utilisateur introuvable"
                    },
                    "422": {
                        "description": "Erreur de validation"
                    }
                }
            }
        },
        "/users/{id}/preferred-language": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Recevoir le langage préféré d’un utilisateur",
                "description": "Analyse les critiques de l'utilisateur pour déterminer son langage le plus utilisé.",
                "operationId": "bd61a3178a50df49392d23c1f975f7e7",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de l'utilisateur",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Langage préféré",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "language": {
                                            "type": "string",
                                            "example": "fr"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Utilisateur introuvable"
                    }
                }
            }
        }
    },
...
}